+++ src/app/(auth)/dashboard/layout.tsx
@@ -98,7 +98,5 @@ export default function DashboardLayout({ children }: { children: React.ReactNod
           </div>
         </div>
       </main>
-      {/* <LoginRequiredModal /> */}
-    </>
   );
 }
